fix(proxy): spawning the proxy in a separate process, same as bridge
